+++ Vidly/Models/Movie.cs
@@ -15,7 +15,7 @@ namespace Vidly.Models
 
         [Required]
         [Display(Name = "Genre")]
-        public byte GenreId { get; set; }
+        public int GenreId { get; set; }
 
         [Required]
         [Display(Name = "Release Date")]
